Remove `Supervisor` and `EvidenceReporter`, as well as associated examples and tests
