+++ light-client/src/builder.rs
@@ -3,7 +3,4 @@
 mod light_client;
 pub use light_client::LightClientBuilder;
 
-mod supervisor;
-pub use supervisor::SupervisorBuilder;
-
 pub mod error;
