+++ light-client/src/instance.rs
@@ -1,0 +1,37 @@
+//! Supervisor and Handle implementation.
+
+use crate::{
+    light_client::LightClient,
+    state::State,
+    verifier::types::{LightBlock, Status},
+};
+
+/// A light client `Instance` packages a `LightClient` together with its `State`.
+#[derive(Debug)]
+pub struct Instance {
+    /// The light client for this instance
+    pub light_client: LightClient,
+
+    /// The state of the light client for this instance
+    pub state: State,
+}
+
+impl Instance {
+    /// Constructs a new instance from the given light client and its state.
+    pub fn new(light_client: LightClient, state: State) -> Self {
+        Self {
+            light_client,
+            state,
+        }
+    }
+
+    /// Get the latest trusted block.
+    pub fn latest_trusted(&self) -> Option<LightBlock> {
+        self.state.light_store.highest(Status::Trusted)
+    }
+
+    /// Trust the given block.
+    pub fn trust_block(&mut self, lb: &LightBlock) {
+        self.state.light_store.update(lb, Status::Trusted);
+    }
+}
